refactor: rename maps api client class
